add deploy_wandb_checkpoints() in new module examples/wrenformer/deploy_wrenformer.py
move make_ensemble_predictions() there from aviary/wrenformer/utils.py
use new deploy_wandb_checkpoints() in examples/wrenformer/{mp,mp_wbm}/use_trained_wrenformer_ensemble.py
